Quick fix bug where first channel could not be >1
It's likely there is a bug in the analyzer, but the DB variables
chanmap and chanmap_start were treated as one, and that would cause
a segfault when chanmap_start got changed to the first channel listed
on chanamp.  So I renamed this variable to start_chanmap, and that seems
to work.