Modify SBSGenericDetector
Eliminate the fChanMapStart adding to the row,col and layer.
Use just as offset when going from internal index of
the fElements vector which is 0 to fNelems-1 for the
Element ID used in the chanmap.

At this stage most useful if detector wants chanmap's
element ID to start at 1 instead of 0.